refactor: migrate plugins to src dir
