+++ api/server.js
@@ -6,7 +6,6 @@ import { knex } from './db/knex-database-connection.js';
 import { config } from './lib/config.js';
 import { authentication } from './lib/infrastructure/authentication.js';
 import { monitoringTools } from './lib/infrastructure/monitoring-tools.js';
-import { plugins } from './lib/infrastructure/plugins/index.js';
 import { routes } from './lib/routes.js';
 import { swaggers } from './lib/swaggers.js';
 import { handleFailAction } from './lib/validate.js';
@@ -30,6 +29,7 @@ import { organizationLearnerRoutes } from './src/prescription/organization-learn
 import { organizationPlaceRoutes } from './src/prescription/organization-place/routes.js';
 import { targetProfileRoutes } from './src/prescription/target-profile/routes.js';
 import { schoolRoutes } from './src/school/routes.js';
+import { plugins } from './src/shared/infrastructure/plugins/index.js';
 import { deserializer } from './src/shared/infrastructure/serializers/jsonapi/deserializer.js';
 // bounded context migration
 import { sharedRoutes } from './src/shared/routes.js';
