+++ api/src/shared/infrastructure/plugins/i18n.js
@@ -5,7 +5,7 @@ const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 const plugin = hapiI18n;
 const options = {
   locales: ['en', 'fr'],
-  directory: __dirname + '/../../../translations',
+  directory: __dirname + '../../../../translations',
   defaultLocale: 'fr',
   queryParameter: 'lang',
   languageHeaderField: 'Accept-Language',
